refactor(types): разделяет типы для полного и неполного объектов предложения; выносит типы комментариев, локации и пользователя в отдельные файлы
