+++ src/types/location.ts
@@ -1,0 +1,5 @@
+export type Location = {
+  latitude: number;
+  longitude: number;
+  zoom: number;
+};
