+++ src/types/user.ts
@@ -1,0 +1,5 @@
+export type User = {
+  name: string;
+  avatarUrl: string;
+  isPro: boolean;
+};
